Move panel event handlers to panel
Fixes #9460
